Replaced some more uint8_t, where it may concern a body-literal index/count

+++ include/vlog/concepts.h
@@ -398,7 +398,7 @@ class Rule {
             return heads[0];
         }
 
-        Literal getHead(uint8_t pos) const {
+        Literal getHead(unsigned pos) const {
             return heads[pos];
         }
 
@@ -416,8 +416,8 @@ class Rule {
             return body;
         }
 
-        uint8_t getNIDBPredicates() const {
-            uint8_t i = 0;
+        unsigned getNIDBPredicates() const {
+            unsigned i = 0;
             for (std::vector<Literal>::const_iterator itr = body.begin(); itr != body.end();
                     ++itr) {
                 if (itr->getPredicate().getType() == IDB) {
@@ -427,8 +427,8 @@ class Rule {
             return i;
         }
 
-        uint8_t getNIDBNotMagicPredicates() const {
-            uint8_t i = 0;
+        unsigned getNIDBNotMagicPredicates() const {
+            unsigned i = 0;
             for (std::vector<Literal>::const_iterator itr = body.begin(); itr != body.end();
                     ++itr) {
                 if (itr->getPredicate().getType() == IDB && !itr->getPredicate().isMagic()) {
@@ -438,8 +438,8 @@ class Rule {
             return i;
         }
 
-        uint8_t getNEDBPredicates() const {
-            uint8_t i = 0;
+        unsigned getNEDBPredicates() const {
+            unsigned i = 0;
             for (std::vector<Literal>::const_iterator itr = body.begin(); itr != body.end();
                     ++itr) {
                 if (itr->getPredicate().getType() == EDB) {
@@ -449,8 +449,8 @@ class Rule {
             return i;
         }
 
-        uint8_t numberOfNegatedLiteralsInBody() {
-            uint8_t result = 0;
+        unsigned numberOfNegatedLiteralsInBody() {
+            unsigned result = 0;
             for (std::vector<Literal>::const_iterator itr = getBody().begin();
                     itr != getBody().end(); ++itr) {
                 if (itr->isNegated()){
